Update projectBackend.py
Changes to authenticate function. It can check if a user is a student or a professor by checking their respective tables in the SQL database. It can also return the object used to check if a user exists to the frontend (theoretically) and then can be used in frontend functions